refactor(*): Change name tags to consts
